Add files via upload
still unfinished ( and ugly) but I wanted to get you guys something to work with in case you got ahead of me and wanted to start connecting to the GUI. So far, I believe I have the connect-to-server dialog set up with pretty much all it needs, along with the chat box, a text box, and a send button for the client side. I'm not very confident that this will be all you guys need to see what I've changed, so let me know if I need to do something different.